Contains various changes related to color schemes and contrast improvements across the Track. Now default theme is Purple-Grey.
