+++ app/build.gradle.kts
@@ -15,8 +15,8 @@ android {
         applicationId = "com.savenko.track"
         minSdk = 26
         targetSdk = 34
-        versionCode = 22
-        versionName = "1.3.3"
+        versionCode = 23
+        versionName = "1.3.3.1"
 
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
         vectorDrawables {
