+++ app/src/main/java/com/savenko/track/data/other/constants/EntitiesDefaults.kt
@@ -3,6 +3,8 @@ package com.savenko.track.data.other.constants
 import com.savenko.track.domain.models.currency.Currency
 import com.savenko.track.domain.models.currency.CurrencyTypes
 import com.savenko.track.presentation.themes.Themes
+import com.savenko.track.presentation.themes.purpleGreyTheme.purpleGreyNew_DarkColorScheme
+import com.savenko.track.presentation.themes.purpleGreyTheme.purpleGreyNew_LightColorScheme
 
 // User defaults
 const val NAME_DEFAULT: String = "User"
@@ -10,12 +12,15 @@ const val LOGIN_COUNT_DEFAULT = 0
 const val BUDGET_DEFAULT: Float = 0.0f
 const val SHOW_PAGE_NAME_DEFAULT: Boolean = true
 const val USE_SYSTEM_THEME_DEFAULT: Boolean = true
-val PREFERABLE_THEME_DEFAULT: Themes = Themes.BlueTheme
 const val NON_CATEGORY_FINANCIALS_DEFAULT = false
 const val GROUPING_CATEGORY_ID_DEFAULT = Int.MIN_VALUE // This value used to compare if user selected any category
 const val EXPENSE_CATEGORY_GROUPING_ID_DEFAULT = 13
 const val INCOME_CATEGORY_GROUPING_ID_DEFAULT = 9
 
+// Color scheme and theme
+val PREFERABLE_THEME_DEFAULT: Themes = Themes.PurpleGreyTheme
+val DEFAULT_LIGHT_COLOR_SCHEME = purpleGreyNew_LightColorScheme
+val DEFAULT_DARK_COLOR_SCHEME = purpleGreyNew_DarkColorScheme
 
 // Entities defaults
 val CURRENCY_DEFAULT =
