+++ app/src/main/java/com/savenko/track/presentation/components/bottomSheet/BottomSheetAmountInput.kt
@@ -75,7 +75,7 @@ fun BottomSheetAmountInput(
                 textStyle = MaterialTheme.typography.titleMedium.copy(
                     fontSize = 54.sp,
                     letterSpacing = 1.3.sp,
-                    color = MaterialTheme.colorScheme.onPrimaryContainer
+                    color = MaterialTheme.colorScheme.primary
                 ),
                 value = currentInputValue.toString(),
                 onValueChange = { newText ->
