--- app/src/main/java/com/savenko/track/presentation/themes/BlueTheme/Color.kt
@@ -1,66 +1,0 @@
-package com.savenko.track.presentation.themes.BlueTheme
-import androidx.compose.ui.graphics.Color
-
-val blue_light_theme_color_primary = Color(0xFF0561A4)
-val blue_light_theme_color_onPrimary = Color(0xFFFFFFFF)
-val md_theme_light_primaryContainer = Color(0xFFD1E4FF)
-val md_theme_light_onPrimaryContainer = Color(0xFF001D36)
-val md_theme_light_secondary = Color(0xFF00629F)
-val md_theme_light_onSecondary = Color(0xFFFFFFFF)
-val md_theme_light_secondaryContainer = Color(0xFFD0E4FF)
-val md_theme_light_onSecondaryContainer = Color(0xFF001D34)
-val md_theme_light_tertiary = Color(0xFF1960A5)
-val md_theme_light_onTertiary = Color(0xFFFFFFFF)
-val md_theme_light_tertiaryContainer = Color(0xFFD4E3FF)
-val md_theme_light_onTertiaryContainer = Color(0xFF001C39)
-val md_theme_light_error = Color(0xFFBA1A1A)
-val md_theme_light_errorContainer = Color(0xFFFFDAD6)
-val md_theme_light_onError = Color(0xFFFFFFFF)
-val md_theme_light_onErrorContainer = Color(0xFF410002)
-val md_theme_light_background = Color(0xFFFAFCFF)
-val md_theme_light_onBackground = Color(0xFF001F2A)
-val md_theme_light_surface = Color(0xFFFAFCFF)
-val md_theme_light_onSurface = Color(0xFF001F2A)
-val md_theme_light_surfaceVariant = Color(0xFFDFE2EB)
-val md_theme_light_onSurfaceVariant = Color(0xFF43474E)
-val md_theme_light_outline = Color(0xFF73777F)
-val md_theme_light_inverseOnSurface = Color(0xFFE1F4FF)
-val md_theme_light_inverseSurface = Color(0xFF003547)
-val md_theme_light_inversePrimary = Color(0xFF9FCAFF)
-val md_theme_light_shadow = Color(0xFF000000)
-val md_theme_light_surfaceTint = Color(0xFF0561A4)
-val md_theme_light_outlineVariant = Color(0xFFC3C6CF)
-val md_theme_light_scrim = Color(0xFF000000)
-
-val blue_dark_theme_color_primary = Color(0xFF9FCAFF)
-val blue_dark_theme_color_onPrimary = Color(0xFF003259)
-val md_theme_dark_primaryContainer = Color(0xFF00497D)
-val md_theme_dark_onPrimaryContainer = Color(0xFFD1E4FF)
-val md_theme_dark_secondary = Color(0xFF9ACBFF)
-val md_theme_dark_onSecondary = Color(0xFF003355)
-val md_theme_dark_secondaryContainer = Color(0xFF004A79)
-val md_theme_dark_onSecondaryContainer = Color(0xFFD0E4FF)
-val md_theme_dark_tertiary = Color(0xFFA4C9FF)
-val md_theme_dark_onTertiary = Color(0xFF00315D)
-val md_theme_dark_tertiaryContainer = Color(0xFF004883)
-val md_theme_dark_onTertiaryContainer = Color(0xFFD4E3FF)
-val md_theme_dark_error = Color(0xFFFFB4AB)
-val md_theme_dark_errorContainer = Color(0xFF93000A)
-val md_theme_dark_onError = Color(0xFF690005)
-val md_theme_dark_onErrorContainer = Color(0xFFFFDAD6)
-val md_theme_dark_background = Color(0xFF001F2A)
-val md_theme_dark_onBackground = Color(0xFFBFE9FF)
-val md_theme_dark_surface = Color(0xFF001F2A)
-val md_theme_dark_onSurface = Color(0xFFBFE9FF)
-val md_theme_dark_surfaceVariant = Color(0xFF43474E)
-val md_theme_dark_onSurfaceVariant = Color(0xFFC3C6CF)
-val md_theme_dark_outline = Color(0xFF8D9199)
-val md_theme_dark_inverseOnSurface = Color(0xFF001F2A)
-val md_theme_dark_inverseSurface = Color(0xFFBFE9FF)
-val md_theme_dark_inversePrimary = Color(0xFF0561A4)
-val md_theme_dark_shadow = Color(0xFF000000)
-val md_theme_dark_surfaceTint = Color(0xFF9FCAFF)
-val md_theme_dark_outlineVariant = Color(0xFF43474E)
-val md_theme_dark_scrim = Color(0xFF000000)
-
-val seed = Color(0xFF324862)
