--- app/src/main/java/com/savenko/track/presentation/themes/PinkTheme/Color.kt
@@ -1,66 +1,0 @@
-package com.savenko.track.presentation.themes.PinkTheme
-import androidx.compose.ui.graphics.Color
-
-val pink_light_theme_primary = Color(0xFFAF2A4D)
-val pink_light_theme_onPrimary = Color(0xFFFFFFFF)
-val md_theme_light_primaryContainer = Color(0xFFFFD9DD)
-val md_theme_light_onPrimaryContainer = Color(0xFF400013)
-val md_theme_light_secondary = Color(0xFF9A4057)
-val md_theme_light_onSecondary = Color(0xFFFFFFFF)
-val md_theme_light_secondaryContainer = Color(0xFFFFD9DF)
-val md_theme_light_onSecondaryContainer = Color(0xFF3F0016)
-val md_theme_light_tertiary = Color(0xFFB71948)
-val md_theme_light_onTertiary = Color(0xFFFFFFFF)
-val md_theme_light_tertiaryContainer = Color(0xFFFFD9DD)
-val md_theme_light_onTertiaryContainer = Color(0xFF400012)
-val md_theme_light_error = Color(0xFFBA1A1A)
-val md_theme_light_errorContainer = Color(0xFFFFDAD6)
-val md_theme_light_onError = Color(0xFFFFFFFF)
-val md_theme_light_onErrorContainer = Color(0xFF410002)
-val md_theme_light_background = Color(0xFFFFFBFF)
-val md_theme_light_onBackground = Color(0xFF3C0028)
-val md_theme_light_surface = Color(0xFFFFFBFF)
-val md_theme_light_onSurface = Color(0xFF3C0028)
-val md_theme_light_surfaceVariant = Color(0xFFF3DDDF)
-val md_theme_light_onSurfaceVariant = Color(0xFF524345)
-val md_theme_light_outline = Color(0xFF847374)
-val md_theme_light_inverseOnSurface = Color(0xFFFFECF2)
-val md_theme_light_inverseSurface = Color(0xFF5B113F)
-val md_theme_light_inversePrimary = Color(0xFFFFB2BD)
-val md_theme_light_shadow = Color(0xFF000000)
-val md_theme_light_surfaceTint = Color(0xFFAF2A4D)
-val md_theme_light_outlineVariant = Color(0xFFD7C1C3)
-val md_theme_light_scrim = Color(0xFF000000)
-
-val pink_dark_theme_primary = Color(0xFFFFB2BD)
-val pink_dark_theme_onPrimary = Color(0xFF670024)
-val md_theme_dark_primaryContainer = Color(0xFF8E0B37)
-val md_theme_dark_onPrimaryContainer = Color(0xFFFFD9DD)
-val md_theme_dark_secondary = Color(0xFFFFB1C0)
-val md_theme_dark_onSecondary = Color(0xFF5F112A)
-val md_theme_dark_secondaryContainer = Color(0xFF7C2940)
-val md_theme_dark_onSecondaryContainer = Color(0xFFFFD9DF)
-val md_theme_dark_tertiary = Color(0xFFFFB2BC)
-val md_theme_dark_onTertiary = Color(0xFF670023)
-val md_theme_dark_tertiaryContainer = Color(0xFF910034)
-val md_theme_dark_onTertiaryContainer = Color(0xFFFFD9DD)
-val md_theme_dark_error = Color(0xFFFFB4AB)
-val md_theme_dark_errorContainer = Color(0xFF93000A)
-val md_theme_dark_onError = Color(0xFF690005)
-val md_theme_dark_onErrorContainer = Color(0xFFFFDAD6)
-val md_theme_dark_background = Color(0xFF3C0028)
-val md_theme_dark_onBackground = Color(0xFFFFD8E8)
-val md_theme_dark_surface = Color(0xFF3C0028)
-val md_theme_dark_onSurface = Color(0xFFFFD8E8)
-val md_theme_dark_surfaceVariant = Color(0xFF524345)
-val md_theme_dark_onSurfaceVariant = Color(0xFFD7C1C3)
-val md_theme_dark_outline = Color(0xFF9F8C8E)
-val md_theme_dark_inverseOnSurface = Color(0xFF3C0028)
-val md_theme_dark_inverseSurface = Color(0xFFFFD8E8)
-val md_theme_dark_inversePrimary = Color(0xFFAF2A4D)
-val md_theme_dark_shadow = Color(0xFF000000)
-val md_theme_dark_surfaceTint = Color(0xFFFFB2BD)
-val md_theme_dark_outlineVariant = Color(0xFF524345)
-val md_theme_dark_scrim = Color(0xFF000000)
-
-val seed = Color(0xFFD24566)
