--- app/src/main/java/com/savenko/track/presentation/themes/PurpleTheme/Color.kt
@@ -1,67 +1,0 @@
-package com.savenko.track.presentation.themes.PurpleTheme
-import androidx.compose.ui.graphics.Color
-
-val purple_light_theme_primary = Color(0xFF5A54A8)
-val purple_light_theme_onPrimary = Color(0xFFFFFFFF)
-val md_theme_light_primaryContainer = Color(0xFFE3DFFF)
-val md_theme_light_onPrimaryContainer = Color(0xFF140463)
-val md_theme_light_secondary = Color(0xFF5A53A8)
-val md_theme_light_onSecondary = Color(0xFFFFFFFF)
-val md_theme_light_secondaryContainer = Color(0xFFE4DFFF)
-val md_theme_light_onSecondaryContainer = Color(0xFF150463)
-val md_theme_light_tertiary = Color(0xFF5B53A8)
-val md_theme_light_onTertiary = Color(0xFFFFFFFF)
-val md_theme_light_tertiaryContainer = Color(0xFFE4DFFF)
-val md_theme_light_onTertiaryContainer = Color(0xFF160362)
-val md_theme_light_error = Color(0xFFBA1A1A)
-val md_theme_light_errorContainer = Color(0xFFFFDAD6)
-val md_theme_light_onError = Color(0xFFFFFFFF)
-val md_theme_light_onErrorContainer = Color(0xFF410002)
-val md_theme_light_background = Color(0xFFFAFCFF)
-val md_theme_light_onBackground = Color(0xFF001F2A)
-val md_theme_light_surface = Color(0xFFFAFCFF)
-val md_theme_light_onSurface = Color(0xFF001F2A)
-val md_theme_light_surfaceVariant = Color(0xFFE5E1EC)
-val md_theme_light_onSurfaceVariant = Color(0xFF47464F)
-val md_theme_light_outline = Color(0xFF787680)
-val md_theme_light_inverseOnSurface = Color(0xFFE1F4FF)
-val md_theme_light_inverseSurface = Color(0xFF003547)
-val md_theme_light_inversePrimary = Color(0xFFC5C0FF)
-val md_theme_light_shadow = Color(0xFF000000)
-val md_theme_light_surfaceTint = Color(0xFF5A54A8)
-val md_theme_light_outlineVariant = Color(0xFFC8C5D0)
-val md_theme_light_scrim = Color(0xFF000000)
-
-val purple_dark_theme_primary = Color(0xFFC5C0FF)
-val purple_dark_theme_onPrimary = Color(0xFF2B2277)
-val md_theme_dark_primaryContainer = Color(0xFF423B8E)
-val md_theme_dark_onPrimaryContainer = Color(0xFFE3DFFF)
-val md_theme_dark_secondary = Color(0xFFC5C0FF)
-val md_theme_dark_onSecondary = Color(0xFF2B2276)
-val md_theme_dark_secondaryContainer = Color(0xFF423B8E)
-val md_theme_dark_onSecondaryContainer = Color(0xFFE4DFFF)
-val md_theme_dark_tertiary = Color(0xFFC6BFFF)
-val md_theme_dark_onTertiary = Color(0xFF2C2276)
-val md_theme_dark_tertiaryContainer = Color(0xFF433B8E)
-val md_theme_dark_onTertiaryContainer = Color(0xFFE4DFFF)
-val md_theme_dark_error = Color(0xFFFFB4AB)
-val md_theme_dark_errorContainer = Color(0xFF93000A)
-val md_theme_dark_onError = Color(0xFF690005)
-val md_theme_dark_onErrorContainer = Color(0xFFFFDAD6)
-val md_theme_dark_background = Color(0xFF001F2A)
-val md_theme_dark_onBackground = Color(0xFFBFE9FF)
-val md_theme_dark_surface = Color(0xFF001F2A)
-val md_theme_dark_onSurface = Color(0xFFBFE9FF)
-val md_theme_dark_surfaceVariant = Color(0xFF47464F)
-val md_theme_dark_onSurfaceVariant = Color(0xFFC8C5D0)
-val md_theme_dark_outline = Color(0xFF928F99)
-val md_theme_dark_inverseOnSurface = Color(0xFF001F2A)
-val md_theme_dark_inverseSurface = Color(0xFFBFE9FF)
-val md_theme_dark_inversePrimary = Color(0xFF5A54A8)
-val md_theme_dark_shadow = Color(0xFF000000)
-val md_theme_dark_surfaceTint = Color(0xFFC5C0FF)
-val md_theme_dark_outlineVariant = Color(0xFF47464F)
-val md_theme_dark_scrim = Color(0xFF000000)
-
-
-val seed = Color(0xFF69659F)
