+++ app/src/main/java/com/savenko/track/presentation/themes/ThemeManager.kt
@@ -2,12 +2,12 @@ package com.savenko.track.presentation.themes
 
 import android.os.Build
 import androidx.compose.runtime.Composable
-import com.savenko.track.presentation.themes.BlueTheme.BlueTheme
-import com.savenko.track.presentation.themes.PinkTheme.PinkTheme
-import com.savenko.track.presentation.themes.PurpleTheme.PurpleTheme
-import com.savenko.track.presentation.themes.RedTheme.RedTheme
-import com.savenko.track.presentation.themes.YellowTheme.YellowTheme
-import com.savenko.track.presentation.themes.DynamicColorTheme.TrackerTheme
+import com.savenko.track.presentation.themes.blueTheme.BlueTheme
+import com.savenko.track.presentation.themes.dynamicColorTheme.TrackerTheme
+import com.savenko.track.presentation.themes.pinkTheme.PinkTheme
+import com.savenko.track.presentation.themes.purpleGreyTheme.PurpleGreyTheme
+import com.savenko.track.presentation.themes.redTheme.RedTheme
+import com.savenko.track.presentation.themes.yellowTheme.YellowTheme
 
 
 @Composable
@@ -34,8 +34,8 @@ fun ThemeManager(
                 }
             }
 
-            Themes.PurpleTheme -> {
-                PurpleTheme {
+            Themes.PurpleGreyTheme -> {
+                PurpleGreyTheme {
                     content()
                 }
             }
