+++ app/src/main/java/com/savenko/track/presentation/themes/Themes.kt
@@ -3,19 +3,18 @@ package com.savenko.track.presentation.themes
 import com.savenko.track.data.other.constants.PREFERABLE_THEME_DEFAULT
 
 sealed class Themes(val name : String){
-    object BlueTheme : Themes("Blue")
-    object PinkTheme : Themes("Pink")
-    object PurpleTheme : Themes("Purple")
-    object RedTheme : Themes("Red")
-    object YellowTheme : Themes("Yellow")
+    data object PurpleGreyTheme : Themes("PurpleGrey")
+    data object BlueTheme : Themes("Blue")
+    data object PinkTheme : Themes("Pink")
+    data object RedTheme : Themes("Red")
+    data object YellowTheme : Themes("Yellow")
 }
-
 
 fun getThemeByName(themeName: String): Themes {
     return when (themeName) {
         "Blue" -> Themes.BlueTheme
         "Pink" -> Themes.PinkTheme
-        "Purple" -> Themes.PurpleTheme
+        "PurpleGrey" -> Themes.PurpleGreyTheme
         "Red" -> Themes.RedTheme
         "Yellow" -> Themes.YellowTheme
         else -> PREFERABLE_THEME_DEFAULT
