+++ app/src/main/java/com/savenko/track/presentation/themes/blueTheme/BlueThemeColorScheme.kt
@@ -1,0 +1,142 @@
+package com.savenko.track.presentation.themes.blueTheme
+
+import androidx.compose.material3.darkColorScheme
+import androidx.compose.material3.lightColorScheme
+import androidx.compose.ui.graphics.Color
+
+val blue_dark_theme_color_primary = Color(0xFF9FCAFF)
+val blue_dark_theme_color_onPrimary = Color(0xFF003259)
+val md_theme_dark_primaryContainer = Color(0xFF00497D)
+val md_theme_dark_onPrimaryContainer = Color(0xFFD1E4FF)
+val md_theme_dark_secondary = Color(0xFF9ACBFF)
+val md_theme_dark_onSecondary = Color(0xFF003355)
+val md_theme_dark_secondaryContainer = Color(0xFF004A79)
+val md_theme_dark_onSecondaryContainer = Color(0xFFD0E4FF)
+val md_theme_dark_tertiary = Color(0xFFA4C9FF)
+val md_theme_dark_onTertiary = Color(0xFF00315D)
+val md_theme_dark_tertiaryContainer = Color(0xFF004883)
+val md_theme_dark_onTertiaryContainer = Color(0xFFD4E3FF)
+val md_theme_dark_error = Color(0xFFFFB4AB)
+val md_theme_dark_errorContainer = Color(0xFF93000A)
+val md_theme_dark_onError = Color(0xFF690005)
+val md_theme_dark_onErrorContainer = Color(0xFFFFDAD6)
+val md_theme_dark_background = Color(0xFF001F2A)
+val md_theme_dark_onBackground = Color(0xFFBFE9FF)
+val md_theme_dark_surface = Color(0xFF001F2A)
+val md_theme_dark_onSurface = Color(0xFFBFE9FF)
+val md_theme_dark_surfaceVariant = Color(0xFF43474E)
+val md_theme_dark_onSurfaceVariant = Color(0xFFC3C6CF)
+val md_theme_dark_outline = Color(0xFF8D9199)
+val md_theme_dark_inverseOnSurface = Color(0xFF001F2A)
+val md_theme_dark_inverseSurface = Color(0xFFBFE9FF)
+val md_theme_dark_inversePrimary = Color(0xFF0561A4)
+val md_theme_dark_surfaceTint = Color(0xFF9FCAFF)
+val md_theme_dark_outlineVariant = Color(0xFF43474E)
+val md_theme_dark_scrim = Color(0xFF000000)
+
+val newBlueTheme_DarkColors = darkColorScheme(
+    primary = blue_dark_theme_color_primary,
+    onPrimary = blue_dark_theme_color_onPrimary,
+    primaryContainer = md_theme_dark_primaryContainer,
+    onPrimaryContainer = md_theme_dark_onPrimaryContainer,
+    secondary = md_theme_dark_secondary,
+    onSecondary = md_theme_dark_onSecondary,
+    secondaryContainer = md_theme_dark_secondaryContainer,
+    onSecondaryContainer = md_theme_dark_onSecondaryContainer,
+    tertiary = md_theme_dark_tertiary,
+    onTertiary = md_theme_dark_onTertiary,
+    tertiaryContainer = md_theme_dark_tertiaryContainer,
+    onTertiaryContainer = md_theme_dark_onTertiaryContainer,
+    error = md_theme_dark_error,
+    errorContainer = md_theme_dark_errorContainer,
+    onError = md_theme_dark_onError,
+    onErrorContainer = md_theme_dark_onErrorContainer,
+    background = md_theme_dark_background,
+    onBackground = md_theme_dark_onBackground,
+    surface = md_theme_dark_surface,
+    onSurface = md_theme_dark_onSurface,
+    surfaceVariant = md_theme_dark_surfaceVariant,
+    onSurfaceVariant = md_theme_dark_onSurfaceVariant,
+    outline = md_theme_dark_outline,
+    inverseOnSurface = md_theme_dark_inverseOnSurface,
+    inverseSurface = md_theme_dark_inverseSurface,
+    inversePrimary = md_theme_dark_inversePrimary,
+    surfaceTint = md_theme_dark_surfaceTint,
+    outlineVariant = md_theme_dark_outlineVariant,
+    scrim = md_theme_dark_scrim,
+)
+
+val blueTheme_primaryLight = Color(0xFF405F91)
+val blueTheme_onPrimaryLight = Color(0xFFFFFFFF)
+val blueTheme_primaryContainerLight = Color(0xFFD6E3FF)
+val blueTheme_onPrimaryContainerLight = Color(0xFF001B3E)
+val blueTheme_secondaryLight = Color(0xFF565F71)
+val blueTheme_onSecondaryLight = Color(0xFFFFFFFF)
+val blueTheme_secondaryContainerLight = Color(0xFFDAE2F9)
+val blueTheme_onSecondaryContainerLight = Color(0xFF131C2B)
+val blueTheme_tertiaryLight = Color(0xFF6F5575)
+val blueTheme_onTertiaryLight = Color(0xFFFFFFFF)
+val blueTheme_tertiaryContainerLight = Color(0xFFF9D8FD)
+val blueTheme_onTertiaryContainerLight = Color(0xFF28132E)
+val blueTheme_errorLight = Color(0xFFBA1A1A)
+val blueTheme_onErrorLight = Color(0xFFFFFFFF)
+val blueTheme_errorContainerLight = Color(0xFFFFDAD6)
+val blueTheme_onErrorContainerLight = Color(0xFF410002)
+val blueTheme_backgroundLight = Color(0xFFF9F9FF)
+val blueTheme_onBackgroundLight = Color(0xFF191C20)
+val blueTheme_surfaceLight = Color(0xFFF9F9FF)
+val blueTheme_onSurfaceLight = Color(0xFF191C20)
+val blueTheme_surfaceVariantLight = Color(0xFFE0E2EC)
+val blueTheme_onSurfaceVariantLight = Color(0xFF44474E)
+val blueTheme_outlineLight = Color(0xFF74777F)
+val blueTheme_outlineVariantLight = Color(0xFFC4C6D0)
+val blueTheme_scrimLight = Color(0xFF000000)
+val blueTheme_inverseSurfaceLight = Color(0xFF2E3036)
+val blueTheme_inverseOnSurfaceLight = Color(0xFFF0F0F7)
+val blueTheme_inversePrimaryLight = Color(0xFFAAC7FF)
+val blueTheme_surfaceDimLight = Color(0xFFD9D9E0)
+val blueTheme_surfaceBrightLight = Color(0xFFF9F9FF)
+val blueTheme_surfaceContainerLowestLight = Color(0xFFFFFFFF)
+val blueTheme_surfaceContainerLowLight = Color(0xFFF3F3FA)
+val blueTheme_surfaceContainerLight = Color(0xFFEDEDF4)
+val blueTheme_surfaceContainerHighLight = Color(0xFFE7E8EE)
+val blueTheme_surfaceContainerHighestLight = Color(0xFFE2E2E9)
+
+val newBlueTheme_LightColors = lightColorScheme(
+    primary = blueTheme_primaryLight,
+    onPrimary = blueTheme_onPrimaryLight,
+    primaryContainer = blueTheme_primaryContainerLight,
+    onPrimaryContainer = blueTheme_onPrimaryContainerLight,
+    secondary = blueTheme_secondaryLight,
+    onSecondary = blueTheme_onSecondaryLight,
+    secondaryContainer = blueTheme_secondaryContainerLight,
+    onSecondaryContainer = blueTheme_onSecondaryContainerLight,
+    tertiary = blueTheme_tertiaryLight,
+    onTertiary = blueTheme_onTertiaryLight,
+    tertiaryContainer = blueTheme_tertiaryContainerLight,
+    onTertiaryContainer = blueTheme_onTertiaryContainerLight,
+    error = blueTheme_errorLight,
+    onError = blueTheme_onErrorLight,
+    errorContainer = blueTheme_errorContainerLight,
+    onErrorContainer = blueTheme_onErrorContainerLight,
+    background = blueTheme_backgroundLight,
+    onBackground = blueTheme_onBackgroundLight,
+    surface = blueTheme_surfaceLight,
+    onSurface = blueTheme_onSurfaceLight,
+    surfaceVariant = blueTheme_surfaceVariantLight,
+    onSurfaceVariant = blueTheme_onSurfaceVariantLight,
+    outline = blueTheme_outlineLight,
+    inverseOnSurface = blueTheme_inverseOnSurfaceLight,
+    inverseSurface = blueTheme_inverseSurfaceLight,
+    inversePrimary = blueTheme_inversePrimaryLight,
+    surfaceTint = blueTheme_primaryLight,
+    outlineVariant = blueTheme_outlineVariantLight,
+    scrim = blueTheme_scrimLight,
+    surfaceDim = blueTheme_surfaceDimLight,
+    surfaceBright = blueTheme_surfaceBrightLight,
+    surfaceContainerLowest = blueTheme_surfaceContainerLowestLight,
+    surfaceContainerLow = blueTheme_surfaceContainerLowLight,
+    surfaceContainer = blueTheme_surfaceContainerLight,
+    surfaceContainerHigh = blueTheme_surfaceContainerHighLight,
+    surfaceContainerHighest = blueTheme_surfaceContainerHighestLight,
+)
