+++ app/src/main/java/com/savenko/track/presentation/themes/blueTheme/Theme.kt
@@ -1,0 +1,34 @@
+package com.savenko.track.presentation.themes.blueTheme
+
+import android.app.Activity
+import androidx.compose.foundation.isSystemInDarkTheme
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.SideEffect
+import androidx.compose.ui.graphics.toArgb
+import androidx.compose.ui.platform.LocalView
+import androidx.core.view.WindowCompat
+
+@Composable
+fun BlueTheme(
+    useDarkTheme: Boolean = isSystemInDarkTheme(),
+    content: @Composable () -> Unit
+) {
+    val colors = if (!useDarkTheme) {
+        newBlueTheme_LightColors
+    } else {
+        newBlueTheme_DarkColors
+    }
+    val view = LocalView.current
+    if (!view.isInEditMode) {
+        SideEffect {
+            val window = (view.context as Activity).window
+            window.statusBarColor = colors.primary.toArgb()
+            WindowCompat.getInsetsController(window, view).isAppearanceLightStatusBars = useDarkTheme
+        }
+    }
+    MaterialTheme(
+        colorScheme = colors,
+        content = content
+    )
+}
