+++ app/src/main/java/com/savenko/track/presentation/themes/dynamicColorTheme/Theme.kt
@@ -1,0 +1,44 @@
+package com.savenko.track.presentation.themes.dynamicColorTheme
+
+import android.app.Activity
+import android.os.Build
+import androidx.compose.foundation.isSystemInDarkTheme
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.dynamicDarkColorScheme
+import androidx.compose.material3.dynamicLightColorScheme
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.SideEffect
+import androidx.compose.ui.graphics.toArgb
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.platform.LocalView
+import androidx.core.view.WindowCompat
+import com.savenko.track.data.other.constants.DEFAULT_DARK_COLOR_SCHEME
+import com.savenko.track.data.other.constants.DEFAULT_LIGHT_COLOR_SCHEME
+
+@Composable
+fun TrackerTheme(
+    darkTheme: Boolean = isSystemInDarkTheme(),
+    dynamicColor: Boolean = true,
+    content: @Composable () -> Unit
+) {
+    val colorScheme = when {
+        dynamicColor && Build.VERSION.SDK_INT >= Build.VERSION_CODES.S -> {
+            val context = LocalContext.current
+            if (darkTheme) dynamicDarkColorScheme(context) else dynamicLightColorScheme(context)
+        }
+        darkTheme -> DEFAULT_DARK_COLOR_SCHEME
+        else -> DEFAULT_LIGHT_COLOR_SCHEME
+    }
+    val view = LocalView.current
+    if (!view.isInEditMode) {
+        SideEffect {
+            val window = (view.context as Activity).window
+            window.statusBarColor = colorScheme.primary.toArgb()
+            WindowCompat.getInsetsController(window, view).isAppearanceLightStatusBars = darkTheme
+        }
+    }
+    MaterialTheme(
+        colorScheme = colorScheme,
+        content = content
+    )
+}
