+++ app/src/main/java/com/savenko/track/presentation/themes/redTheme/Color.kt
@@ -1,0 +1,67 @@
+package com.savenko.track.presentation.themes.redTheme
+import androidx.compose.ui.graphics.Color
+
+val red_light_theme_primary = Color(0xFFAD302E)
+val red_light_theme_onPrimary = Color(0xFFFFFFFF)
+val md_theme_light_primaryContainer = Color(0xFFFFDAD6)
+val md_theme_light_onPrimaryContainer = Color(0xFF410003)
+val md_theme_light_secondary = Color(0xFF9C4141)
+val md_theme_light_onSecondary = Color(0xFFFFFFFF)
+val md_theme_light_secondaryContainer = Color(0xFFFFDAD8)
+val md_theme_light_onSecondaryContainer = Color(0xFF410006)
+val md_theme_light_tertiary = Color(0xFFAC3232)
+val md_theme_light_onTertiary = Color(0xFFFFFFFF)
+val md_theme_light_tertiaryContainer = Color(0xFFFFDAD7)
+val md_theme_light_onTertiaryContainer = Color(0xFF410004)
+val md_theme_light_error = Color(0xFFBA1A1A)
+val md_theme_light_errorContainer = Color(0xFFFFDAD6)
+val md_theme_light_onError = Color(0xFFFFFFFF)
+val md_theme_light_onErrorContainer = Color(0xFF410002)
+val md_theme_light_background = Color(0xFFFFFBFF)
+val md_theme_light_onBackground = Color(0xFF410005)
+val md_theme_light_surface = Color(0xFFFFFBFF)
+val md_theme_light_onSurface = Color(0xFF410005)
+val md_theme_light_surfaceVariant = Color(0xFFF5DDDB)
+val md_theme_light_onSurfaceVariant = Color(0xFF534342)
+val md_theme_light_outline = Color(0xFF857371)
+val md_theme_light_inverseOnSurface = Color(0xFFFFEDEB)
+val md_theme_light_inverseSurface = Color(0xFF5F1416)
+val md_theme_light_inversePrimary = Color(0xFFFFB3AD)
+val md_theme_light_shadow = Color(0xFF000000)
+val md_theme_light_surfaceTint = Color(0xFFAD302E)
+val md_theme_light_outlineVariant = Color(0xFFD8C2BF)
+val md_theme_light_scrim = Color(0xFF000000)
+
+val red_dark_theme_primary = Color(0xFFFFB3AD)
+val red_dark_theme_onPrimary = Color(0xFF680009)
+val md_theme_dark_primaryContainer = Color(0xFF8C171A)
+val md_theme_dark_onPrimaryContainer = Color(0xFFFFDAD6)
+val md_theme_dark_secondary = Color(0xFFFFB3B0)
+val md_theme_dark_onSecondary = Color(0xFF5F1317)
+val md_theme_dark_secondaryContainer = Color(0xFF7E2A2B)
+val md_theme_dark_onSecondaryContainer = Color(0xFFFFDAD8)
+val md_theme_dark_tertiary = Color(0xFFFFB3AE)
+val md_theme_dark_onTertiary = Color(0xFF68000C)
+val md_theme_dark_tertiaryContainer = Color(0xFF8B191D)
+val md_theme_dark_onTertiaryContainer = Color(0xFFFFDAD7)
+val md_theme_dark_error = Color(0xFFFFB4AB)
+val md_theme_dark_errorContainer = Color(0xFF93000A)
+val md_theme_dark_onError = Color(0xFF690005)
+val md_theme_dark_onErrorContainer = Color(0xFFFFDAD6)
+val md_theme_dark_background = Color(0xFF410005)
+val md_theme_dark_onBackground = Color(0xFFFFDAD7)
+val md_theme_dark_surface = Color(0xFF410005)
+val md_theme_dark_onSurface = Color(0xFFFFDAD7)
+val md_theme_dark_surfaceVariant = Color(0xFF534342)
+val md_theme_dark_onSurfaceVariant = Color(0xFFD8C2BF)
+val md_theme_dark_outline = Color(0xFFA08C8A)
+val md_theme_dark_inverseOnSurface = Color(0xFF410005)
+val md_theme_dark_inverseSurface = Color(0xFFFFDAD7)
+val md_theme_dark_inversePrimary = Color(0xFFAD302E)
+val md_theme_dark_shadow = Color(0xFF000000)
+val md_theme_dark_surfaceTint = Color(0xFFFFB3AD)
+val md_theme_dark_outlineVariant = Color(0xFF534342)
+val md_theme_dark_scrim = Color(0xFF000000)
+
+
+val seed = Color(0xFFCF4944)
