+++ app/src/main/java/com/savenko/track/presentation/themes/yellowTheme/Color.kt
@@ -1,0 +1,67 @@
+package com.savenko.track.presentation.themes.yellowTheme
+import androidx.compose.ui.graphics.Color
+
+val yellow_light_theme_primary = Color(0xFF6A5F00)
+val yellow_light_theme_onPrimary = Color(0xFFFFFFFF)
+val md_theme_light_primaryContainer = Color(0xFFF9E44B)
+val md_theme_light_onPrimaryContainer = Color(0xFF201C00)
+val md_theme_light_secondary = Color(0xFF686000)
+val md_theme_light_onSecondary = Color(0xFFFFFFFF)
+val md_theme_light_secondaryContainer = Color(0xFFF4E569)
+val md_theme_light_onSecondaryContainer = Color(0xFF1F1C00)
+val md_theme_light_tertiary = Color(0xFF695F00)
+val md_theme_light_onTertiary = Color(0xFFFFFFFF)
+val md_theme_light_tertiaryContainer = Color(0xFFF8E54D)
+val md_theme_light_onTertiaryContainer = Color(0xFF201C00)
+val md_theme_light_error = Color(0xFFBA1A1A)
+val md_theme_light_errorContainer = Color(0xFFFFDAD6)
+val md_theme_light_onError = Color(0xFFFFFFFF)
+val md_theme_light_onErrorContainer = Color(0xFF410002)
+val md_theme_light_background = Color(0xFFFFFBFF)
+val md_theme_light_onBackground = Color(0xFF1D1C16)
+val md_theme_light_surface = Color(0xFFFFFBFF)
+val md_theme_light_onSurface = Color(0xFF1D1C16)
+val md_theme_light_surfaceVariant = Color(0xFFE8E2D0)
+val md_theme_light_onSurfaceVariant = Color(0xFF4A4739)
+val md_theme_light_outline = Color(0xFF7B7768)
+val md_theme_light_inverseOnSurface = Color(0xFFF5F0E7)
+val md_theme_light_inverseSurface = Color(0xFF32302A)
+val md_theme_light_inversePrimary = Color(0xFFDBC82F)
+val md_theme_light_shadow = Color(0xFF000000)
+val md_theme_light_surfaceTint = Color(0xFF6A5F00)
+val md_theme_light_outlineVariant = Color(0xFFCCC6B5)
+val md_theme_light_scrim = Color(0xFF000000)
+
+val yellow_dark_theme_primary = Color(0xFFDBC82F)
+val yellow_dark_theme_onPrimary = Color(0xFF373100)
+val md_theme_dark_primaryContainer = Color(0xFF504700)
+val md_theme_dark_onPrimaryContainer = Color(0xFFF9E44B)
+val md_theme_dark_secondary = Color(0xFFD6C950)
+val md_theme_dark_onSecondary = Color(0xFF363100)
+val md_theme_dark_secondaryContainer = Color(0xFF4E4800)
+val md_theme_dark_onSecondaryContainer = Color(0xFFF4E569)
+val md_theme_dark_tertiary = Color(0xFFDAC831)
+val md_theme_dark_onTertiary = Color(0xFF373100)
+val md_theme_dark_tertiaryContainer = Color(0xFF4F4700)
+val md_theme_dark_onTertiaryContainer = Color(0xFFF8E54D)
+val md_theme_dark_error = Color(0xFFFFB4AB)
+val md_theme_dark_errorContainer = Color(0xFF93000A)
+val md_theme_dark_onError = Color(0xFF690005)
+val md_theme_dark_onErrorContainer = Color(0xFFFFDAD6)
+val md_theme_dark_background = Color(0xFF1D1C16)
+val md_theme_dark_onBackground = Color(0xFFE7E2D9)
+val md_theme_dark_surface = Color(0xFF1D1C16)
+val md_theme_dark_onSurface = Color(0xFFE7E2D9)
+val md_theme_dark_surfaceVariant = Color(0xFF4A4739)
+val md_theme_dark_onSurfaceVariant = Color(0xFFCCC6B5)
+val md_theme_dark_outline = Color(0xFF959181)
+val md_theme_dark_inverseOnSurface = Color(0xFF1D1C16)
+val md_theme_dark_inverseSurface = Color(0xFFE7E2D9)
+val md_theme_dark_inversePrimary = Color(0xFF6A5F00)
+val md_theme_dark_shadow = Color(0xFF000000)
+val md_theme_dark_surfaceTint = Color(0xFFDBC82F)
+val md_theme_dark_outlineVariant = Color(0xFF4A4739)
+val md_theme_dark_scrim = Color(0xFF000000)
+
+
+val seed = Color(0xFFDCC930)
